A06guncellenecek-A07tam-B06tam

+++ src/main/java/com/cookerytech/service/BrandService.java
@@ -1,6 +1,7 @@
 package com.cookerytech.service;
 
 import com.cookerytech.domain.Brand;
+import com.cookerytech.domain.Product;
 import com.cookerytech.domain.Role;
 import com.cookerytech.domain.enums.RoleType;
 import com.cookerytech.dto.BrandDTO;
@@ -24,6 +25,7 @@ public class BrandService {
 
     private final BrandRepository brandRepository;
     private final BrandMapper brandMapper;
+    private final ProductService productService;
 
     private final UserService userService;
 
@@ -31,10 +33,11 @@ public class BrandService {
 
 
 
-
-    public BrandService(BrandRepository brandRepository, BrandMapper brandMapper, UserService userService, RoleService roleService) {
+    public BrandService(BrandRepository brandRepository, BrandMapper brandMapper, ProductService productService,
+                        UserService userService,RoleService roleService) {
         this.brandRepository = brandRepository;
         this.brandMapper = brandMapper;
+        this.productService = productService;
         this.userService = userService;
         this.roleService = roleService;
     }
@@ -82,6 +85,12 @@ public class BrandService {
 
         if (brand.getBuiltIn()){
             throw new BadRequestException(ErrorMessage.NOT_PERMITTED_METHOD_MESSAGE);
+        }
+
+       List<Product> productList=productService.getProductByBrandId(id);
+
+        if(productList.size()>0){
+            throw new BadRequestException(ErrorMessage.BRAND_CAN_NOT_DELETED);
         }
 
         brandRepository.delete(brand);
